+++ starlite/enums.py
@@ -4,11 +4,13 @@ from enum import Enum
 class HttpMethod(str, Enum):
     """An Enum for HTTP methods."""
 
+    DELETE = "DELETE"
     GET = "GET"
+    HEAD = "HEAD"
+    OPTIONS = "OPTIONS"
+    PATCH = "PATCH"
     POST = "POST"
     PUT = "PUT"
-    PATCH = "PATCH"
-    DELETE = "DELETE"
 
 
 class MediaType(str, Enum):
